chore: updates workflow to enable npm trusted publishing
NPM package publishing will be migrated to use Trusted Publishing (via OIDC).

To ensure security and restrict access, we will configure the OIDC token request to only be available to jobs running in the dedicated 'release' environment. This environment will be a required condition for obtaining the necessary NPM token.

Note: `workflow_dispatch:` will be added for testing the publishing. It will be removed in a follow up PR, which will also remove the currently configured token.